refactor: make app independent of hardhat
